Move processScaffoldFiles to scaffold collection, and move scaffold collection to Operations/ScaffoldOpCollection
